+++ cm/cmind/automation.py
@@ -148,6 +148,9 @@ class Automation:
                                        and just load meta to match UID/alias
 
 
+            (skip_index_search) (bool): If True, skip indexing
+            (force_index_add) (bool): If True, force index add (for reindexing)
+
         Returns: 
             (CM return dict):
 
@@ -161,6 +164,9 @@ class Automation:
         import copy
 
         console = i.get('out') == 'con'
+
+        skip_index_search = i.get('skip_index_search', False)
+        force_index_add = i.get('force_index_add', False)
 
         lst = []
 
@@ -185,7 +191,7 @@ class Automation:
         # Second optional object in a list is a repo
         parsed_automation = i.get('parsed_automation',[])
 
-        auto_name = parsed_automation[0] if len(parsed_automation)>0 else ('','')
+        automation_name = parsed_automation[0] if len(parsed_automation)>0 else ('','')
 
         # Get parsed artifact
         parsed_artifact = i.get('parsed_artifact',[])
@@ -223,13 +229,50 @@ class Automation:
             if add_repo:
                  pruned_repos.append(repo)
 
+
+        # Check index and bypass search
+        if not skip_index_search:
+            r = self.cmind.index.find(automation_name, artifact_obj, artifact_repo, pruned_repos, and_tags, no_tags)
+            if r['return']==0:
+                artifacts = r.get('list',[])
+                if len(artifacts)>0:
+                    lst = []
+
+                    repo_extra_info = self.cmind.repos.extra_info
+
+                    for artifact in artifacts:
+                        artifact_object = Artifact(cmind = self.cmind, path = artifact['path'])
+
+                        for j in [0,1]:
+                            x = artifact['repo'][j]
+                            if x in repo_extra_info:
+                                # Add extra info to artifact about the repo
+                                artifact_object.repo_path = repo_extra_info[x].path
+                                artifact_object.repo_meta = repo_extra_info[x].meta
+
+                        r = artifact_object.load(ignore_inheritance = ignore_inheritance)
+                        if r['return']>0: return r
+
+                        lst.append(artifact_object)
+
+                        # Output to console if forced
+                        if console:
+                            print (artifact_object.path)
+
+                    return {'return':0, 'list':lst}
+        
+        
+        
+
+        
         # Iterate over pruned repositories
         for repo in pruned_repos:
             path_repo = repo.path_with_prefix
 
             repo_meta = copy.deepcopy(repo.meta)
 
-            automations = os.listdir(path_repo)
+            # May or may not be automation
+            automations = sorted(os.listdir(path_repo))
 
             for automation in automations:
                 path_automations = os.path.join(path_repo, automation)
@@ -240,7 +283,8 @@ class Automation:
 
                     path_artifacts = os.path.join(path_repo, path_automations)
 
-                    artifacts = os.listdir(path_artifacts)
+                    # Potential CM artifacts or just some directories
+                    artifacts = sorted(os.listdir(path_artifacts))
 
                     # Check if artifact is first to get meta about automation UID/alias
                     first_artifact = True
@@ -249,6 +293,7 @@ class Automation:
                         path_artifact = os.path.join(path_artifacts, artifact)
 
                         if os.path.isdir(path_artifact):
+                            
                             # Check if has CM meta to make sure that it's a CM object
                             path_artifact_meta = os.path.join(path_artifact, self.cmind.cfg['file_cmeta'])
 
@@ -266,6 +311,7 @@ class Automation:
                                 # Force no inheritance if first artifact just to check automation UID and alias
                                 tmp_ignore_inheritance = True if first_artifact else ignore_inheritance
 
+                                # Force no inheritance to get basic info about aliases and UIDs
                                 r = artifact_object.load(ignore_inheritance = tmp_ignore_inheritance)
                                 if r['return']>0: return r
 
@@ -282,8 +328,8 @@ class Automation:
                                     # Need to check if automation matches
                                     r = utils.match_objects(uid = meta.get('automation_uid'), 
                                                             alias = meta.get('automation_alias'),
-                                                            uid2 = auto_name[1],
-                                                            alias2 = auto_name[0],
+                                                            uid2 = automation_name[1],
+                                                            alias2 = automation_name[0],
                                                             more_strict = True)
                                     if r['return']>0: return r
 
@@ -304,20 +350,18 @@ class Automation:
                                         r = artifact_object.load(ignore_inheritance = ignore_inheritance)
                                         if r['return']>0: return r
 
-                                    tags_in_meta = meta.get('tags',[])
+                                    meta = artifact_object.meta
 
-                                    if len(and_tags)>0:
-                                        if not all(t in tags_in_meta for t in and_tags):
-                                            continue
+                                    # Index
+                                    if self.cmind.use_index or force_index_add:
+                                        r=self.cmind.index.add(meta, path_artifact, 
+                                                                     (repo.meta['alias'], 
+                                                                      repo.meta['uid']))
+                                        # Ignore index output to continue working if issues
 
-                                    if len(no_tags)>0:
-                                        skip = False
-                                        for t in no_tags:
-                                            if t in tags_in_meta:
-                                                skip = True
-                                                break
-                                        if skip:
-                                            continue
+                                    # Check if tags match
+                                    if not utils.tags_matched(meta.get('tags',[]), and_tags, no_tags):
+                                        continue
 
                                     lst.append(artifact_object)
 
@@ -327,7 +371,6 @@ class Automation:
 
                                 if first_artifact:
                                     first_artifact = False
-
 
         return {'return':0, 'list':lst}
 
@@ -374,7 +417,7 @@ class Automation:
 
         parsed_automation = i.get('parsed_automation',[])
 
-        auto_name = parsed_automation[0] if len(parsed_automation)>0 else ('','')
+        automation_name = parsed_automation[0] if len(parsed_automation)>0 else ('','')
 
         # Get parsed artifact
         parsed_artifact = i.get('parsed_artifact',[])
@@ -406,7 +449,7 @@ class Automation:
         repo_path = repo.path_with_prefix
 
         # Check automation (if exists)
-        automation_path = os.path.join(repo_path, auto_name[0]) if auto_name[0]!='' else os.path.join(repo_path, auto_name[1])
+        automation_path = os.path.join(repo_path, automation_name[0]) if automation_name[0]!='' else os.path.join(repo_path, automation_name[1])
 
         if not os.path.isdir(automation_path):
             os.makedirs(automation_path)
@@ -436,8 +479,8 @@ class Automation:
 
         if meta.get('alias','')=='': meta['alias']=artifact_obj[0]
         if meta.get('uid','')=='': meta['uid']=artifact_obj[1]
-        if meta.get('automation_alias','')=='': meta['automation_alias']=auto_name[0]
-        if meta.get('automation_uid','')=='': meta['automation_uid']=auto_name[1]
+        if meta.get('automation_alias','')=='': meta['automation_alias']=automation_name[0]
+        if meta.get('automation_uid','')=='': meta['automation_uid']=automation_name[1]
 
         existing_tags = meta.get('tags',[])
         if len(tags)>0: 
@@ -451,6 +494,14 @@ class Automation:
             r = utils.save_json(meta_path_json, meta=meta)
         if r['return']>0: return r
 
+        # Index
+        if self.cmind.use_index:
+            r=self.cmind.index.add(meta, obj_path, 
+                                   (repo.meta['alias'], repo.meta['uid']),
+                                   update = True)
+            # Ignore index output to continue working if issues
+
+        
         if console:
             print ('Added CM object at {}'.format(obj_path))
 
@@ -537,6 +588,13 @@ class Automation:
             else:
                 shutil.rmtree(path_to_artifact)
 
+            # Index
+            if self.cmind.use_index:
+                r=self.cmind.index.add(artifact.meta, artifact.path, 
+                                       (artifact.repo_meta['alias'], artifact.repo_meta['uid']),
+                                       update = True, delete = True)
+                # Ignore index output to continue working if issues
+            
             if console:
                 print ('    Deleted!')
 
@@ -708,6 +766,13 @@ class Automation:
 
             r = artifact.update(meta, replace = replace, append_lists = not replace_lists, tags = meta_tags)
 
+            # Index
+            if self.cmind.use_index:
+                r=self.cmind.index.add(artifact.meta, artifact.path, 
+                                       (artifact.repo_meta['alias'], artifact.repo_meta['uid']),
+                                       update = True)
+                # Ignore index output to continue working if issues
+
             # Output if console
             if console:
                 print ('Updated {}'.format(artifact.path))
@@ -775,6 +840,7 @@ class Automation:
 
         # Check target repo
         target_artifact_repo = target_artifact[1] if len(target_artifact)>1 else None
+        target_artifact_repo_obj = None
 
         target_repo_path = None
 
@@ -791,10 +857,18 @@ class Automation:
             elif len(target_repo_list) >1:
                 return {'return':1, 'error':'more than 1 target repo found "{}"'.format(target_artifact_repo)}
 
-            target_repo_path = os.path.abspath(target_repo_list[0].path_with_prefix)
+            target_artifact_repo_obj = target_repo_list[0]
+            target_repo_path = os.path.abspath(target_artifact_repo_obj.path_with_prefix)
 
         # Updating artifacts
         for artifact in lst:
+
+            # Index
+            if self.cmind.use_index:
+                r=self.cmind.index.add(artifact.meta, artifact.path, 
+                                       (artifact.repo_meta['alias'], artifact.repo_meta['uid']),
+                                       update = True, delete = True)
+                # Ignore index output to continue working if issues
 
             artifact_path = os.path.abspath(artifact.path)
 
@@ -859,6 +933,19 @@ class Automation:
                     r = artifact.update({})
                     if r['return'] >0: return r
 
+            # Index
+            if self.cmind.use_index:
+                if target_artifact_repo_obj is not None:
+                    tmp_repo_index = (target_artifact_repo_obj.meta['alias'], 
+                                      target_artifact_repo_obj.meta['uid'])
+                else:
+                   tmp_repo_index = (artifact.repo_meta['alias'], 
+                                     artifact.repo_meta['uid'])
+                
+                r=self.cmind.index.add(artifact.meta, artifact.path, 
+                                       tmp_repo_index, update = True)
+                # Ignore index output to continue working if issues
+
         return {'return':0, 'list':lst}
 
     ############################################################
@@ -922,11 +1009,11 @@ class Automation:
 
         # Check target repo
         target_artifact_repo = target_artifact[1] if len(target_artifact)>1 else None
+        target_artifact_repo_obj = None
 
         target_repo_path = None
 
         if target_artifact_repo is not None:
-            print (target_artifact_repo)
             r = self.cmind.access({'action':'search',
                                    'automation':'repo',
                                    'artifact': utils.assemble_cm_object2(target_artifact_repo)})
@@ -939,7 +1026,8 @@ class Automation:
             elif len(target_repo_list) >1:
                 return {'return':1, 'error':'more than 1 target repo found "{}"'.format(target_artifact_repo)}
 
-            target_repo_path = os.path.abspath(target_repo_list[0].path_with_prefix)
+            target_artifact_repo_obj = target_repo_list[0]
+            target_repo_path = os.path.abspath(target_artifact_repo_obj.path_with_prefix)
 
         # Updating artifacts
         for artifact in lst:
@@ -992,6 +1080,7 @@ class Automation:
             # If only yaml, update yaml and not json
             meta_path_yaml = os.path.join(new_artifact_path, self.cmind.cfg['file_cmeta']+'.yaml')
             meta_path_json = os.path.join(new_artifact_path, self.cmind.cfg['file_cmeta']+'.json')
+
             if os.path.isfile(meta_path_yaml) and not os.path.isfile(meta_path_json):
                 update_meta={'alias':artifact_meta['alias'],
                              'uid':artifact_meta['uid']}
@@ -1002,6 +1091,21 @@ class Automation:
             else:
                 r = artifact.update({})
                 if r['return'] >0: return r
+
+            # Index added artifact
+            if self.cmind.use_index:
+                # Old
+                if target_artifact_repo_obj is not None:
+                    tmp_repo_index = (target_artifact_repo_obj.meta['alias'], 
+                                      target_artifact_repo_obj.meta['uid'])                    
+                else:
+                    tmp_repo_index = (artifact.repo_meta['alias'], 
+                                      artifact.repo_meta['uid'])
+                                      
+                r=self.cmind.index.add(artifact.meta, artifact.path, 
+                                       tmp_repo_index, update = True)
+                # Ignore index output to continue working if issues
+
 
         return {'return':0, 'list':lst}